syscall: import syscall crate
Import the new `r-linux-syscall` crate. This crate provides direct
access to linux system calls, as well as prototypes for all known parts
of the kernel ABI.

This crate is far from complete. It contains the stubs to directly
invoke system calls, but it lacks the prototypes of the different parts
of the kernel ABI. However, this will be a huge task to add, so the
crate will definitely be used before all bits are there. In fact, there
is little aim in making this crate complete. There are just too many
old, outdated, and archaic parts of the kernel API that are just too
much effort to import, and too little users.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>