ci: drop rustfmt
The rustfmt tool does not accept variable input and adapt, but always
enforces a uniform formatting, including line-breaks. This works
contrary to our style, where we like to break lines even if they do not
exceed the 100ch limit.

As long as rustfmt does not gain any features to configure this, we will
disable it.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>